ArtisanaWebsite/skillhub-africa/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
django-redis==5.4.0
boto3==1.34.0
django-storages==1.14.2
celery==5.3.4
twilio==8.11.0
textblob==0.17.1
django-cors-headers==4.3.1
